Restructure source and build path
