+++ example/index.html
@@ -5,10 +5,10 @@
     <link rel="stylesheet" href="https://elect.in.th/assets/typography.css" />
 
     <script type="module">
-      import '../dist/staging-banner.js'
-      import '../dist/sharer.js'
-      import '../dist/footer.js'
-      import '../dist/navbar.js'
+      import '../components/staging-banner.js';
+      import '../components/sharer.js';
+      import '../components/footer.js';
+      import '../components/navbar.js';
     </script>
   </head>
