+++ package.json
@@ -2,7 +2,6 @@
   "name": "@electinth/common",
   "version": "1.2.3",
   "description": "ELECT's reusable component repository",
-  "main": "dist/index.js",
   "repository": "https://github.com/electinth/common.git",
   "author": "Th1nkK1D <witheep@gmail.com>",
   "license": "MIT",
@@ -11,7 +10,7 @@
     "build-storybook": "build-storybook",
     "new": "hygen component new",
     "lint": "eslint . --fix",
-    "build:components": "rm -rf dist/* && node esbuild.mjs",
+    "build:components": "rm -rf dist/* && node src/utils/esbuild.mjs",
     "build:assets": "cp -rf assets dist && postcss dist/assets/*.css --replace",
     "build": "npm run build:components && npm run build:assets"
   },
@@ -52,9 +51,9 @@
     }
   },
   "lint-staged": {
-    "*.{js,svelte,ts.json}": "eslint --cache --fix"
+    "**/*": "prettier --write --ignore-unknown"
   },
   "files": [
-    "dist/**/*"
+    "components/**/*"
   ]
 }
