+++ src/components/footer/companies.ts
