+++ src/components/footer/elect.ts
