+++ src/components/footer/footer.stories.svelte
@@ -1,6 +1,6 @@
 <script>
   import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
-  import Footer from '../../components/footer.svelte';
+  import Footer from './footer.wc.svelte';
 </script>
 
 <Meta title="Components/footer" component={Footer} />
