+++ src/components/footer/social.ts
