+++ src/components/sharer/sharer.wc.svelte
