+++ src/utils/esbuild.mjs
@@ -3,16 +3,22 @@ import { build } from 'esbuild';
 import esbuildSvelte from 'esbuild-svelte';
 import sveltePreprocess from 'svelte-preprocess';
 import { preprocess as windiCss } from 'svelte-windicss-preprocess';
-import transformStyleForWebComponent from './utils/style.mjs';
+import transformStyleForWebComponent from './style.mjs';
 
 (async () => {
-  const components = await glob('./components/**/*.svelte');
+  const components = await glob('src/components/**/*.wc.svelte');
 
-  components.forEach((component) => {
+  components.forEach((componentPath) => {
+    const fileName = componentPath
+      .split('/')
+      .reverse()[0]
+      .replace('.wc.svelte', '.js');
+
     build({
-      entryPoints: [component],
+      entryPoints: [componentPath],
+      outfile: `./components/${fileName}`,
       bundle: true,
-      outdir: './dist',
+      outdir: '',
       plugins: [
         esbuildSvelte({
           preprocess: [
